+++ src/cacheout/cache.py
@@ -203,32 +203,18 @@ class Cache(object):
 
         return value
 
-    def get_many(self, keys, default=None):
-        """Return many cache values as a ``dict`` of key/value pairs.
+    def get_many(self, iteratee, default=None):
+        """Return many cache values as a ``dict`` of key/value pairs filtered
+        by an `iteratee` that can be one of:
 
-        Args:
-            keys (list): List of cache keys.
-            default (mixed, optional): Value to return if key doesn't exist.
-                Defaults to ``None``.
-
-        Returns:
-            dict
-        """
-        return {key: self.get(key, default=default) for key in keys}
-
-    def get_many_by(self, iteratee, default=None):
-        """Like :meth:`get_many` except that it accepts an `iteratee` that can
-        be used to filter the cache keys. The `iteratee` can be one of:
-
+        - ``list`` - List of cache keys.
         - ``str`` - Search string that supports Unix shell-style wildcards.
         - ``re.compile()`` - Compiled regular expression.
         - ``function`` - Function that returns whether a key matches. Invoked
           with ``iteratee(key)``.
-        - ``list``` - List of keys to match on. This is equivalent to calling
-          ``Cache.get_many(iteratee)``.
 
         Args:
-            iteratee (str|Pattern|callable|list): Iteratee to filter by.
+            iteratee (list|str|Pattern|callable): Iteratee to filter by.
             default (mixed, optional): Value to return if key doesn't exist.
                 Defaults to ``None``.
 
@@ -236,7 +222,8 @@ class Cache(object):
             dict
         """
         with self._lock:
-            return self.get_many(self._filter(iteratee), default=default)
+            return {key: self.get(key, default=default)
+                    for key in self._filter(iteratee)}
 
     def add(self, key, value, ttl=None):
         """Add cache key/value if it doesn't already exist. Essentially, this
@@ -333,41 +320,30 @@ class Cache(object):
 
         return count
 
-    def delete_many(self, keys):
-        """Delete multiple cache keys at once.
+    def delete_many(self, iteratee):
+        """Delete multiple cache keys at once filtered by an `iteratee` that
+        can be one of:
+
+        - ``list`` - List of cache keys.
+        - ``str`` - Search string that supports Unix shell-style wildcards.
+        - ``re.compile()`` - Compiled regular expression.
+        - ``function`` - Function that returns whether a key matches. Invoked
+          with ``iteratee(key)``.
 
         Args:
-            keys (list): List of cache keys.
+            iteratee (list|str|Pattern|callable): Iteratee to filter by.
 
         Returns:
             int: Number of cache keys deleted.
         """
         count = 0
 
-        for key in keys:
-            count += self.delete(key)
+        with self._lock:
+            keys = list(self._filter(iteratee))
+            for key in keys:
+                count += self.delete(key)
 
         return count
-
-    def delete_many_by(self, iteratee):
-        """Like :meth:`delete_many` except that it accepts an `iteratee` that
-        can be used to filter the cache keys. The `iteratee` can be one of:
-
-        - ``str`` - Search string that supports Unix shell-style wildcards.
-        - ``re.compile()`` - Compiled regular expression.
-        - ``function`` - Function that returns whether a key matches. Invoked
-          with ``iteratee(key)``.
-        - ``list``` - List of keys to match on. This is equivalent to calling
-          ``Cache.get_many(iteratee)``.
-
-        Args:
-            iteratee (str|Pattern|callable|list): Iteratee to filter by.
-
-        Returns:
-            int: Number of cache keys deleted.
-        """
-        with self._lock:
-            return self.delete_many(list(self._filter(iteratee)))
 
     def delete_expired(self):
         """Delete expired cache keys and return number of entries deleted.
@@ -478,6 +454,9 @@ class Cache(object):
         return (key, value)
 
     def _filter(self, iteratee):
+        # By default, we'll filter against cache storage.
+        target = self._cache
+
         if isinstance(iteratee, str):
             filter_by = re.compile(fnmatch.translate(iteratee)).match
         elif isinstance(iteratee, re._pattern_type):
@@ -485,10 +464,18 @@ class Cache(object):
         elif callable(iteratee):
             filter_by = iteratee
         else:
-            def filter_by(key):
-                return key in iteratee
+            # We're assuming that iteratee is now an iterable that we want to
+            # filter cache keys by. We can optimize the filtering by making
+            # the filter target be the list of keys and checking whether those
+            # keys are in the cache. I.e. we'll iterate over the iteratee keys
+            # can check if the key is in the cache as opposed to iterating over
+            # the cache and checking if a key is in iteratee keys.
+            target = iteratee
 
-        return filter(filter_by, self._cache)
+            def filter_by(key):
+                return key in self._cache
+
+        return filter(filter_by, target)
 
     def memoize(self, *, ttl=None, typed=False):
         """Decorator that wraps a function with a memoizing callable.